Changed functions to take size_t length arguments.

+++ include/cmd_interface.h
@@ -72,7 +72,7 @@ typedef enum
  *
  * @returns 0 on success, 1 on error
  */
-  int msg_arg(char *pipe, int pipe_len, int cmd, char *arg, int arg_len);
+  int msg_arg(char *pipe, size_t pipe_len, int cmd, char *arg, int arg_len);
 
 /**
  * @brief Creates and sends message then listens for a list of responses on the pipe provided.
@@ -83,5 +83,5 @@ typedef enum
  *
  * @returns 0 on success, 1 on error
  */
-  int msg_list(char *pipe, int pipe_len, int cmd);
+  int msg_list(char *pipe, size_t pipe_len, int cmd);
 #endif
